backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
sqlmodel==0.0.14
SQLAlchemy<2.1,>=2.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.36
imageio-ffmpeg==0.4.9
python-multipart
psycopg2-binary==2.9.7
